averaged colors and started improving zoom
Now the bottom level gif takes the colors of the top level gif and layers it onto itself, making it look like the top gif once it zooms out enough. Started to implement percentage zoom to make the zoom more consistent instead of it quickly zooming in the beginning and slowly zooming in the end.